Support connecting to prestarted openvpn
For "PRESTARTED" connections, no openvpn is started but a connection
to a already active management interface is attempted. The disconnect
button puts the connection on hold and connect button releases it.
Once tagged as "PRESTARTED" a config remains in that state
even if MI disconnects. This means such configs connot be reconnected
until the management interface comes up possibly by a retsart of
the service. On exit, the openvpn instance is left running in their state.

TODO: have a config parameter to indicate a connection be treated
as PRESTARTED. Currently presence of management directive serves this
purpose.

Signed-off-by: Selva Nair <selva.nair@gmail.com>