Support for sending a message from server to GUI using echo
Process four new echo commands to construct messages to be
displayed to the user:

  echo msg message-text
  echo msg-n message-text
  echo msg-window message-title
  echo msg-notify message-title

Note: All rules of push and echo processing apply and determine
what is received as echo commands by the GUI. In addition,
'url-encoded' characters (% followed by two hex digits) are
decoded and displayed.

The message is constructed in the GUI by concatenating the text
specified in one or more "echo msg text" or "echo msg-n text"
commands. In case of "echo msg text" the previous text, if any,
is appended with a new line. An empty text in this case will
just add a new line.

The message ends and gets displayed when one of the following
are receieved:

echo msg-window title
echo msg-notify title

where "title" becomes the title of the message window. In case of
msg-window, a modeless window shows the mssage, in the latter case
a notification balloon is shown.

Example: when pushed from the server:

push "echo msg I say let the world go to hell%2C"
push "echo msg I must have my cup of tea."
push "echo msg-window Notes from the underground"

will display a modeless window with title
"Notes from the underground" and a two line body
--
I say let the world go to hell,
I must have my cup of tea.
--

Note that the message itself is not quoted in the above examples
and so it relies on the server's option-parser combining
individual words into a space separated string. Number of words
on a line is limited by the maximum number of parameters allowed
on a command line (15). This limitation may be avoided by quoting
the text that follows.

The comma character is not allowed in pushed strings, so
it has to be sent encoded as %2C as shown above.
Such encoding of arbitrary bytes is suppored. For example,
newlines may be embedded this (needs to be %0D%0A on windows)
although the same may be achieved in a cleaner way using
multiple msg directives.

An example with embedded spaces and multiple lines concatenated
without a new line in between (note use of single quotes)

push "echo msg I swear to you gentlemen%2C that to be"
push "echo msg-n ' overly conscious is a sickness%2C ' "
push "echo msg-n a real%2C thorough sickness."
push "echo msg-notify Quote of the Day"

will show up as a notofication that displays for an
OS-dependent interval as:

--
Quote of the Day

I swear to you gentlemen, that to be overly conscious
is a sickness, a real, thorough sickness.
--

where the location of the line break is automatically determined
by the notification API and is OS version-dependent.

Commands like "echo msg ..." in the config file are also
processed the same way. It gets displayed when the GUI connects
to the management interface and receives all pending echo.
Pushed message(s) get displayed when the client daemon
processes push-reply and passes on echo directives to the
GUI.

- msg-window : displayed when starting from a disconnected state
               not when reconnecting by SIGUSR1 or SIGHUP

- msg-notify : displayed during first conenction as well as
               reconnections -- ie. whenever echo or pushed
               echo are received.

In both cases repeated messages are suppressed for a period
(24 h by default) -- set by popup_mute_interval.

It may be set from command-line as --popup_mute_interval 12
(for 12 hours). Use 0 for no muting.

TODO:
- make the repeated message suppression run-time configurable
- save msg history when GUI exits and reload on start

Signed-off-by: Selva Nair <selva.nair@gmail.com>