Retry on management timeout instead of aborting
In  some cases the service may take a while to startup openvpn.exe
causing management timeout and an  OpenVPN process left behind
when it eventually starts up (Trac 1050).

Fix by retrying on timeout unless an error is received from
the service or the OpenVPN process.

- User can abort by pressing disconnect
- Call StopOpenVPN() before abort so that OpenVPN daemon
  is not left running in case it starts up later
- Eliminate the timedout state that is no longer used
- If OpenManagement() fails (unlikely to happen), show an error
  and call StopOpenVPN() before exit

A "retrying.." message is logged on to the status window every
15 seconds.

See Trac: #1050

Signed-off-by: Selva Nair <selva.nair@gmail.com>