Handle >REMOTE query from daemon
Add support for selecting the server (remote) from
the GUI. Requires --management-query-remote in the
config file. This option is not added by the GUI.

A list of all available remotes are presented to the
user with the proposed remote as selected. The list
could be sorted alphabetically by clicking on any
column header.

If there are no duplicate remote entries, the selected
remote can be passed to the daemon in one round trip
(in about ~1 second).  Otherwise it can take a few
round trips, or as much as O(n) round trips in some
pathological cases, where n is the number of remote entries.

Some remote entries could be disabled in the core
(CE_DISABLED). Currently the GUI does have this
information, and selecting one such entry would lead to
a failed connection attempt.

TODO:
  auto-submit the previous selection or default after a delay
  in case of no user interaction.

Signed-off-by: Selva Nair <selva.nair@gmail.com>