WIP: Add a global option to always prompt for OTP
This adds a checkbox to the general settings menu
to always prompt for OTP when prompting for pasword
even if no static-challenge request is received from
the core. The prompt text may be localized. The response
and password are concatenated to a single string and send
in plain to the management interface.

This is suggested as an alternative to OpenVPN core
supporting an extension to the static-challenge format.
The latter is preferred over this approach.

Signed-off-by: Selva Nair <selva.nair@gmail.com>